Move named_data_types.py into ddt.py
